Only put values in array for checkbox state on checkbox click
- Previously, the SparseBooleanArray representing  the all apps' screen
checkboxes and their corresponding state has new values added to it
whenever the checkbox state changes via OnCheckedChangeListener.
However, for a certain number of installed apps, scrolling the add apps
list screen triggers the bind() method, which also triggers a reset of
the checkbox state and a reset of the corresponding SparseBooleanArray
value.
- Thus, instead of saving the checkbox state via
OnCheckedChangeListener, it is only saved when the checkbox is clicked.
This is done by calling the OnClickListener callback on the checkbox.
- Closes #4
Signed-off-by: Alvin Dizon <alvin.dizon91@gmail.com>